Bill C32, 2010
<http://www2.parl.gc.ca/HousePublications/Publication.aspx?Language=E&Parl=40&Ses=3&Mode=1&Pub=Bill&Doc=C-32_1&File=75>

There are some clairifying transitional provisions and coming into force
provisions at the end of the bill that are really notes, and not part of the
diffable suggestions.

This patch was produced from the official text of the bill by
Stephen Paul Weber, but may contain small omissions or transcriptions errors.
Please report any problems to <c32@singpolyma.net>